Fix useLoadMore composable
