+++ src/pages/search/index.vue
@@ -7,6 +7,7 @@ import { useMeta, defineComponent } from '@nuxtjs/composition-api'
 
 import { propTypes } from '~/pages/search/search-page.types'
 import { useLoadMore } from '~/composables/use-load-more'
+import { ALL_MEDIA } from '~/constants/media'
 
 import VAllResultsGrid from '~/components/VAllResultsGrid/VAllResultsGrid.vue'
 
@@ -17,7 +18,7 @@ const SearchIndex = defineComponent({
   setup(props) {
     useMeta({ title: `${props.searchTerm} | Openverse` })
 
-    const { canLoadMore, onLoadMore } = useLoadMore(props)
+    const { canLoadMore, onLoadMore } = useLoadMore(props, ALL_MEDIA)
     return { canLoadMore, onLoadMore }
   },
   head: {
